Added DVP

+++ src/simModel/Output.java
@@ -12,8 +12,8 @@ public class Output {
 	int totalNumberT12Contracts;
 	int contractsT34satisfied;
 	int totalNumberT34Contracts;
-	// GAComment: Why should first Cost variables are integers?
-	int fixedTotalCost;
+	// TODO: GAComment: Why should first Cost variables are integers?
+	double fixedTotalCost;
 	double overtimeCost;
 	double averageDailyCost;
 	
@@ -40,7 +40,7 @@ public class Output {
 		}
 		return 0;
 	}
-	// GAcomment: why not provide method to compute the averageDailyCost at the end of the observation interval
+	// TODO: GAcomment: why not provide method to compute the averageDailyCost at the end of the observation interval
 	//            as specified in the CM Output specs.
 
 }
